requirements.txt
# Core Dependencies for AI Red Masking Workflow
torch>=2.8.0
torchvision>=0.23.0
Pillow>=11.0.0
numpy>=2.0.0

# AI Detection Models
segment_anything>=1.0
timm>=1.0.19
transformers>=4.0.0

# Additional Dependencies
addict>=2.4.0
yapf>=0.43.0
platformdirs>=4.3.0

# ComfyUI Dependencies (will be installed from ComfyUI/requirements.txt)
# - Additional packages will be installed during setup process